Adds support for a pair argument.  Also addresses multibyte offset issue. (#11)
The Python encode allows the passing of two arguments - a sequence and a pair - that are subsequently combined and encoded.  This makes sense in certain prompt/completion or question/answer workflows.  This adds support for an optional pair.

Additionally, when I was looking at the Rust implementation, I noted a distinction between the code that the Python binding was calling (encode_char_offsets) and what the Ruby binding was calling (encode) on the underlying Rust tokenizer.  This leads to a difference in how offsets are calculated for strings that contain multibyte characters in the tokenizer vocabulary.  I updated the binding to call a function matching the Python version.